% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Paschold2012}
\alias{Paschold2012}
\title{Paschold et. al. (2012) data}
\format{A \code{data.frame} with 634,496 rows and 6 variables:
\describe{
  \item{GeneID}{gene identifier}
  \item{genotype}{B73, Mo17, B73xMo17, or Mo17xB73}
  \item{replicate}{1-4}
  \item{flow_cell}{1-2}
  \item{B73}{read count attributed to B73 genome}
  \item{Mo17}{read count attributed to Mo17}
  \item{total}{read count for that gene}
}}
\usage{
Paschold2012
}
\description{
A data set containing RNAseq counts for 4 replicates of each of 4 genotypes:
B73, Mo17, B73xMo17, and Mo17xB73.
}
\details{
Provides RNA sequencing experiment used in Paschold et. al. (2012). 
The data consist of RNAseq counts for 39,656 "genes" for 4 replicates for two 
parental varieties of maize and their two crosses: Mo17, B73, Mo17xB73, and 
B73xMo17. In addition to the count for each replicate-sample combination, the 
data also contain the counts by parental genotype. 

The data have been lengthed so that each sample is in a unique row with B73 and 
Mo17 indicating the count for each gene attributed to B73 and Mo17 genome and 
the total column indicating the total count for that gene. As some genes can not be 
attributed to a particular genome, the sum of the B73 and Mo17 columns may not 
equal the total column, but the sum should never be greater than the B73 and Mo17
columns. Non-zero counts in the B73 column for the Mo17 genotype (and the reverse)
indicate false positive reads.
}
\references{
Paschold, A., Jia, Y., Marcon, C., Lund, S., Larson, N.B., Yeh, C.T., Ossowski, S., 
Lanz, C., Nettleton, D., Schnable, P.S. and Hochholdinger, F., 2012. Complementation 
contributes to transcriptome complexity in maize (Zea mays L.) hybrids relative to 
their inbred parents. Genome research, 22(12), pp.2445-2454.
}
\keyword{datasets}

